Moved functions away from index.js
