+++ src/Watchable.js
@@ -1,5 +1,3 @@
-import { set } from './index';
-
 export default class Watchable {
   constructor(source, prop, parent) {
     this.path = `${source.__path}.${prop}`;
@@ -13,6 +11,22 @@ export default class Watchable {
   }
 
   set(value) {
-    return set(this.path.split('.'), this.parent(), value);
+    const shallow = this.getShallowSource();
+    return (shallowSource[prop] = value);
+  }
+
+  getShallowSource() {
+    const pathNesting = this.getPathNesting();
+    const source = this.parent();
+    let shallowSource = source;
+    for (let pathNestingLevel of pathNesting) {
+      shallowSource = shallowSource[pathNestingLevel];
+    }
+    return shallowSource;
+  }
+
+  getPathNesting() {
+    const path = this.path.split('.');
+    return path.slice(1, path.length - 1);
   }
 }
